Slow but Steady Progress
Added an AngledRaycast function that can, well, raycast at an angle. I was going to use this in determining which nodes the player can travel to next, but I realized that it might be easier to integrate path generation into the map generation code, which I'll work on starting tomorrow.